renamed test code and written some git event tests
fixed a data race with watcher state
